# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EquationValidationInputMode = typeof EquationValidationInputMode[keyof typeof EquationValidationInputMode];


export const EquationValidationInputMode = {
  auto: 'auto',
  function: 'function',
  parametric: 'parametric',
  implicit: 'implicit',
  implicit3d: 'implicit3d',
  polar: 'polar',
  vector: 'vector',
  piecewise: 'piecewise',
  points: 'points',
} as const;

export interface EquationValidationInput {
  /** @minLength 1 */
  equation: string;
  mode: EquationValidationInputMode;
}

export type EquationValidationResultMode = typeof EquationValidationResultMode[keyof typeof EquationValidationResultMode];


export const EquationValidationResultMode = {
  function: 'function',
  parametric: 'parametric',
  implicit: 'implicit',
  implicit3d: 'implicit3d',
  polar: 'polar',
  vector: 'vector',
  piecewise: 'piecewise',
  points: 'points',
} as const;

export type EquationValidationResultVerificationStatus = typeof EquationValidationResultVerificationStatus[keyof typeof EquationValidationResultVerificationStatus];


export const EquationValidationResultVerificationStatus = {
  invalid: 'invalid',
  static: 'static',
  animatable: 'animatable',
} as const;

export interface EquationValidationResult {
  valid: boolean;
  mode: EquationValidationResultMode;
  animatable: boolean;
  hasTimeVariable: boolean;
  supports2dFallback: boolean;
  verificationStatus: EquationValidationResultVerificationStatus;
  verificationMessage: string;
  /** @nullable */
  normalized: string | null;
  /** @nullable */
  error: string | null;
  suggestions: string[];
  variables: string[];
}

export interface ErrorResponse {
  error: string;
}

export type FeedbackCategory = typeof FeedbackCategory[keyof typeof FeedbackCategory];


export const FeedbackCategory = {
  bug: 'bug',
  suggestion: 'suggestion',
} as const;

export interface Feedback {
  id: number;
  name: string;
  timestamp: string;
  category: FeedbackCategory;
  content: string;
}

export type FeedbackInputCategory = typeof FeedbackInputCategory[keyof typeof FeedbackInputCategory];


export const FeedbackInputCategory = {
  bug: 'bug',
  suggestion: 'suggestion',
} as const;

export interface FeedbackInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  category: FeedbackInputCategory;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}