# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackCategory } from './feedbackCategory';

export interface Feedback {
  id: number;
  name: string;
  timestamp: Date;
  category: FeedbackCategory;
  content: string;
}