# lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackInputCategory } from './feedbackInputCategory';

export interface FeedbackInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  category: FeedbackInputCategory;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}